Avoid sinon import in exported lib files (#3685)
